Don't use Checkboxes to configure signing and encryption in SignEncryptWidget
The "Encrypt for others" UI by default has an enabled checkbox, but no keys selected. This could make a user believe that the data is being encrypted
(since the checkbox is set) when it actually isn't. The checkbox is also redundant, since it doesn't change anything by itself.

To have a consistent UI, the checkboxes for "encrypt for me" and "sign" are also replaced with a "do not sign/encrypt" option in the dropdown.

GnuPG-Bug-ID: 6485