Ensure that menu is shown after clicked item is set as current item
We'll update the actions (enabled/disabled) whenever the current item
changes. Therefore, we have to make sure that the clicked item is set
as current item before the menu is shown. Since Qt shows the menu on
pressed we cannot set it on the Actions button on creation. Instead we
set and show the menu ourselves when the button is pressed and we have
updated the current item.

GnuPG-bug-id: 7018