Ensure correct state of card slot actions
Instead of using the identical actions for all card slots use individual
"proxy" actions for the card slots. This avoids problems with the state
of the actions, e.g. if some action needs to be disabled for one card
slot but enabled for another card slot. The proxy actions copy text,
tool tip and icon from the shared actions and forward the triggered
signal to them.

GnuPG-bug-id: 7018