Fix location of default context menu with Copy action
For item views we have to use mapToGlobal on the viewport because the
coordinates are relative to the viewport. This could be seen in the
Edit Group dialog where the context menu was shown a bit shifted to the
top (by the height of the column header).

(cherry picked from commit 2ff87097d1b4851d56268b00c3f7db456f4dd8af)