Use the key ID instead of the short key ID for filenames
The short key ID is the last 4 octets of the key ID. For V5 keys the
key ID is the first 8 octets of the fingerprint, so that the short key
ID is octets 5-8 of the fingerprint. This means that for V5 keys the
short key ID isn't easily matched with the fingerprint.

GnuPG-bug-id: 7301