Fix double-& in desktop action names for folder actions
The double-& were added in 2011 with comment "Avoid ampersand
being used for an accelerator". I guess this misinterpretation of a
single & has been fixed in the meantime. The "Sign & Encrypt File"
action also has just a single & and this doesn't confuse Dolphin
when it shows accelerators on pressing Alt. The deskop file
specification also doesn't say anything about & in the Name entry.

BUG: 491674