Show "view-refresh" icon as progress indicator
The "process-working-kde.mng" movie is gone since ages. Replace it with
a static "view-refresh" icon (that's also used by KBusyIndicatorWidget)
and remove the long useless (and not deprecated) KIconLoader::loadMovie
call.

When we have too much time we can use the animated KBusyIndicatorWidget
to indicate background processing for this rarely used dialog. Until
then the static icon is good enough.