Remove support for setting long obsolete gpg-agent option
With GnuPG 2.0.21 the option "allow-mark-trusted" was replaced by the
option "no-allow-mark-trusted". That was in 2013. In Gpg4win this change
had been used even earlier (since 2011-01-24). Time to remove this
setting from Kleopatra which was anyway hidden for all newer versions
of GnuPG. People still using perhistoric versions of GnuPG can still set
the option in gpg-agent.conf.

GnuPG-bug-id: 7350