Treat empty result without error as canceled secret key export job
The gpgme job doesn't report a canceled operation (because gpg doesn't
report it). If the export job doesn't return any data although no error
is reported then it should be safe to assume that the operation was
canceled. Entering no password or a wrong password does result in an
appropriate error.

GnuPG-bug-id: 6091