Fix crash in debug dialog
When running a command that doesn't terminate (e.g., "gpg") and closing the window, the process for the command
is killed, causing QProcess::errorOccurred to be emitted. We're connecting to this signal to change UI elements, which
have already been deleted at this point. This patch fixes that by changing the connections to have the UI components as receiver,
which makes them disconnect when the objects are destroyed

GnuPG-Bug-ID: 7331
(cherry picked from commit de346048d83a4280fdeb76f330eb8fedbbd90536)