Update action button when updating the list of card keys
When the certificate belonging to a card key is imported from the card
or deleted locally then we need to update the action button of the
corresponding list item. We only need to do this if there's only one
action (because the menu actions are updated whenever the menu is opened).

GnuPG-bug-id: 7413