Learn card certificates without relying on gpgsm --learn-card
gpgsm --learn-card doesn't work well if multiple cards are inserted
because it simply learns the certificates of the current card which may
not be the card we want to learn. Therefore, we import the card
certificates manually similar to how gpgsm does this in combination
with gpg-agent's "LEARN --send" command.

GnuPG-bug-id: 7379