Draw overlay widget with semi-transparent background
This way everything drawn by the base widget (behind the overlay widget)
is dimmed. Disabling the base widget didn't have the desired effect
because QTreeWidgetItems with foreground color are still drawn with this
color while items without foreground color are drawn with the "disabled"
color if the QTreeWidget is disabled.

GnuPG-bug-id: 7431