Import missing card certificates when cards are reloaded
When the smartcards are (re)loaded then setCard is called by the slot
reacting to the cardChanged signal. In this case we want to (re)import
missing card certificates from the card. On the other hand, we don't
want to reimport the card certificates if they were deleted (which
triggers an update of the certificate list) or after the validation
(which is basically just a listing of the already known card certificates).

GnuPG-bug-id: 7433