Fix crash when output directory for decryption doesn't exist
The code previously just asserted that the directory exists.
Instead, bring over the equivalent code from DecryptVerifyFilesDialog by putting it into a helper function.
This also shows appropriate errors when the path is empty, not a directory, etc.
To be able to "cancel" going to the next page, WizardPage::onNext was changed to return a bool indicating whether the wizard should actually continue.

GnuPG-Bug-ID: 7439