Fix clipboard actions on wayland
We can't reliably monitor the clipboard in the background on wayland, so the existing logic
for enabling/disabling the actions depending on the clipboard's data doesn't work.

Instead, we always enable these actions when on wayland and show an error if the content isn't right.

The "import" command already mostly has this logic, but required some changes to always show the dialog.