Avoid cursor jumping to end of line edit when user edits search term
We have a signal-slot loop to keep the search term in sync between
different components. By comparing the "new" value with the current
value we break the loop and avoid that the cursor jumps to the end of
the text in the input field when the user types into the input field.

GnuPG-bug-id: 7510