Rework pinentry parenting on wayland
The previous approach didn't work well when multiple kleopatra windows are involved, since we always exported the main window.
Instead, we now listen to the focused window changing and always export this one. KWindowSystem makes sure that we don't actually
export the same window twice and that windows that are deleted are cleaned up properly, so this shouldn't leak any resources.