Revert "Ensure correct high-contrast palette on Windows"
This reverts most of the changes made with commit
a2324cd49a861eb98eeb1266ef8ce4815ac465c7.
It does not revert the removal of the one-time changes of the
application palette in case of active high-contrast mode.

For gpg4win we have now patches for qtbase which ensure correct colors
for high-contrast mode. Upstreaming is pending.

GnuPG-bug-id: 7594