Use a null icon for an empty input field
There's no "unavailable" status icon (but we want to use status icons
instead of emblem icons). Using no icon is anyway nicer because an
empty input field isn't really important. The null icon takes the same
space as the actual icons so that there's no jumping when the user enters
a recipient.

GnuPG-bug-id: 7588