Don't crash with failed assert if expired or revoked user ID is selected
We allow the users to select an expired or revoked user ID via the
selection dialog, i.e. expired and revoked are no longer invalid status.
The crash only occurs if Q_ASSERT is active, i.e. with debug builds or
with release builds that are build with QT_FORCE_ASSERTS.