Handle command line options of Qt and KAboutData if app is already running
We have to handle the command line options ourselves because
QCommandLineParser::process() or KAboutData::processCommandLine()
would exit the running Kleopatra process after handling --help,
--version, --author, etc.

Qt doesn't provide the text for --help-all. We show an error instead.
For --version, --author and --license we open the About dialog to
keep it simple.

GnuPG-bug-id: 7686