Open certificate details when pressing enter in certificate list
The slot connected to QAbstractItemView::activated previously did effectively nothing. Replace it with the logic for double-clicking
to make the view also react to pressing return or enter.

A different option would be to add a shortcut to the Details command. This doesn't work well with allowing both Enter and Return to be used, though.
A minor drawback of not using an explicit shortcut is that the shortcut isn't registered with the other shortcuts and thus can't be discovered as easily.

(cherry picked from commit e93d3247e1d0ea5886e32b7b08fc2ffac850639c)