Store last used folders persistently
Kleopatra did already cache the last used folders for several workflows
like import/export, encryption/decryption, etc., but only in memory.
This mechanism is changed so that the last used folders survive a restart
of Kleopatra by storing them in the state config.

Additionally, the default location is changed from the user's home
directory to the user's Documents folder. Especially, on Windows the
user's home directory is not a common, easy accessible location anymore.

GnuPG-bug-id: 7802