Use prefilled CN/EMAIL if no CN/EMAIL is configured
If CN_prefill/EMAIL_prefill is true then CN/EMAIL should be prefilled with
values found locally. Therefore we need to use the prefilled values as
fallback for the values read from the configuration. Otherwise the
prefilled values are overwritten with empty strings if no fixed CN/EMAIL
is configured. This fixes a regression introduced with the rewrite.

GnuPG-bug-id: 6117