Prevent NVDA from speaking "X.509 Directory Services" twice
Qt reports the title of a group box as accessible name for the direct
children of the group box (similar to how labels work) unless the children
report their own accessible name. In case of the directory services editor
the list and the buttons are contained in a plain QWidget which doesn't
report its own accessible name so that the group box title was used as
last resort. This can be avoided by explicitly setting a non-empty
accessible name for the DirectoryServicesWidget.

GnuPG-bug-id: 7850