Set pubkey URL to a single space if setting to empty string fails
Some OpenPGP smartcards don't allow setting the pubkey URL attribute to
an empty value. gpg-card sets a single space when the user wants to
clear the pubkey URL. We now do the same if setting the pubkey URL to
an empty string failed.

Additionally, the pubkey URL attribute is now percent-plus-encoded as
done by gpg-card. This will prevent problems with URLs containing the
plus character or spaces.

GnuPG-bug-id: 7881