Don't make the dialog larger than necessary
We don't need to add 20 pixels to the calculated height. And for the
width we want to add the left and right margin (instead of a magic 20),
but only to the content width of the expander. The size hint of the
dialog already contains the margins.

GnuPG-bug-id: 6117
(backport commit libkleo 29fdf1f8748d1f45be79362a07d9accdf20a92d4)